requirements-txt.txt
# ============================================================
# ROADY CONSTRUCTION - DÉPENDANCES PYTHON
# pip install -r requirements.txt
# ============================================================

# ============================================================
# FRAMEWORK WEB
# ============================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.1
python-multipart==0.0.6
httpx==0.26.0
websockets==12.0

# ============================================================
# BASE DE DONNÉES
# ============================================================
sqlalchemy==2.0.25
asyncpg==0.29.0              # PostgreSQL async driver
psycopg2-binary==2.9.9       # PostgreSQL sync driver
alembic==1.13.1              # Migrations
geoalchemy2==0.14.3          # PostGIS support

# ============================================================
# CACHE & QUEUE
# ============================================================
redis==5.0.1
celery==5.3.6
flower==2.0.1                 # Celery monitoring

# ============================================================
# AUTHENTIFICATION & SÉCURITÉ
# ============================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
cryptography==41.0.7

# ============================================================
# VALIDATION & SERIALIZATION
# ============================================================
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================================
# LLM & AI
# ============================================================
anthropic==0.18.1            # Claude API
openai==1.12.0               # GPT API
google-generativeai==0.3.2   # Gemini API
tiktoken==0.5.2              # Token counting
langchain==0.1.5             # LLM framework (optional)

# ============================================================
# NOTIFICATIONS
# ============================================================
firebase-admin==6.3.0        # Push notifications
twilio==8.12.0               # SMS

# ============================================================
# EMAIL
# ============================================================
aiosmtplib==3.0.1
jinja2==3.1.3                # Email templates

# ============================================================
# FICHIERS & STORAGE
# ============================================================
boto3==1.34.34               # AWS S3
google-cloud-storage==2.14.0 # GCS
python-magic==0.4.27         # File type detection
pillow==10.2.0               # Image processing
pypdf==4.0.1                 # PDF processing

# ============================================================
# EXPORT & DOCUMENTS
# ============================================================
openpyxl==3.1.2              # Excel
reportlab==4.0.9             # PDF generation
weasyprint==60.2             # HTML to PDF
xlsxwriter==3.1.9            # Excel writer

# ============================================================
# MONITORING & LOGGING
# ============================================================
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.2
structlog==24.1.0
python-json-logger==2.0.7

# ============================================================
# UTILITAIRES
# ============================================================
python-dotenv==1.0.1
pyyaml==6.0.1
orjson==3.9.12               # Fast JSON
python-dateutil==2.8.2
pytz==2024.1
slugify==0.0.1
shortuuid==1.0.11
bcrypt==4.1.2

# ============================================================
# GÉOLOCALISATION
# ============================================================
geopy==2.4.1
shapely==2.0.2

# ============================================================
# TESTING
# ============================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0                # Pour tests async
factory-boy==3.3.0           # Test factories
faker==22.5.0                # Fake data

# ============================================================
# DEV TOOLS
# ============================================================
black==24.1.1                # Formatter
isort==5.13.2                # Import sorting
flake8==7.0.0                # Linter
mypy==1.8.0                  # Type checking
pre-commit==3.6.0

# ============================================================
# DOCUMENTATION
# ============================================================
mkdocs==1.5.3
mkdocs-material==9.5.6
